feat: generate cv with user info
