+++ package.json
@@ -16,7 +16,9 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "commitizen": "^4.3.0",
     "emailjs-com": "^3.2.0",
+    "file-saver": "^2.0.5",
     "firebase": "^9.22.0",
+    "jspdf": "^2.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-jss": "^10.10.0",
