+++ src/components/content/DetailsContent.jsx
@@ -1,163 +1,15 @@
-import React, { useState, useEffect } from "react";
-import { createUseStyles } from "react-jss";
+import { useState, useEffect } from "react";
 import { doc, setDoc } from "firebase/firestore";
 import { firestore } from "../../utils/Firestore";
 import { getAuth } from "firebase/auth";
-import Header from "../header/Header";
 import { useUserDetails } from "../getDetails/getDetails";
 import "../../fonts.css";
-import businessCard from "../../assets/business-card.png";
-import resume from "../../assets/resume.png";
+import WorkHistoryForm from "./WorkHistoryForm";
+import PersonalDetailsForm from "./PersonalDetailsForm";
+import handleGeneratePDF from "./HandleGeneratePDF";
+import useStyles from "./useStylesDetails";
 
 const auth = getAuth();
-
-const useStyles = createUseStyles({
-  main: {
-    display: "flex",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#f5f5f5",
-    color: "black",
-    height: "auto",
-    width: "60%",
-    margin: "0 auto",
-    "& h2": {
-      fontFamily: "Delicious Handrawn, cursive",
-      fontSize: "2rem",
-      color: "black",
-    },
-    "& h3": {
-      fontFamily: "Raleway, sans-serif",
-    },
-    "& p": {
-      fontFamily: "Raleway, sans-serif",
-    },
-  },
-  body: {
-    color: "black",
-    "& h2": {
-      fontFamily: "Delicious Handrawn, cursive",
-      fontSize: "2rem",
-      color: "black",
-    },
-    "& h3": {
-      fontFamily: "Raleway, sans-serif",
-    },
-    "& p": {
-      fontFamily: "Raleway, sans-serif",
-    },
-  },
-  form: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    color: "black",
-    maxWidth: "600px",
-    margin: "0 auto",
-    padding: "1rem",
-    "& h2": {
-      fontFamily: "Delicious Handrawn, cursive",
-      fontSize: "2rem",
-    },
-    "& label": {
-      fontFamily: "Raleway, sans-serif",
-    },
-    "& h3": {
-      fontFamily: "Raleway, sans-serif",
-    },
-  },
-  titleimg: {
-    display: "flex",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    width: "60%",
-    "& img": {
-      maxWidth: "70px",
-      paddingLeft: "1rem",
-    },
-  },
-  input: {
-    color: "black",
-    backgroundColor: "#f0f0f0",
-    borderColor: "#666",
-    borderRadius: "10px",
-    width: "100%",
-    padding: "0.5rem",
-    margin: "0.5rem 0",
-    border: "1px solid #ddd",
-  },
-  skillinput: {
-    color: "black",
-    backgroundColor: "#f0f0f0",
-    borderColor: "#666",
-    borderRadius: "10px",
-    width: "60%",
-    padding: "0.5rem",
-    margin: "0.5rem 0",
-    border: "1px solid #ddd",
-  },
-  submitButton: {
-    border: "none",
-    background: "none",
-    cursor: "pointer",
-    backgroundColor: "var(--primary-colour)",
-    color: "white",
-    borderRadius: "0.5rem",
-    marginTop: "1rem",
-    width: "100%",
-    padding: "1rem",
-    fontSize: "1.2rem",
-    fontWeight: "bold",
-    "&:hover": {
-      backgroundColor: "var(--secondary-colour)",
-    },
-  },
-  namesdiv: {
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    "& input": {
-      width: "48%",
-    },
-  },
-  closebutton: {
-    border: "none",
-    background: "none",
-    cursor: "pointer",
-    transition: "0.5s all ease-in-out",
-    "&:hover": {
-      color: "red",
-      transition: "0.5s all ease-in-out",
-      transform: "scale(1.1)",
-    },
-  },
-  checkboxLabel: {
-    display: "flex",
-    alignItems: "center",
-    fontFamily: "Raleway, sans-serif",
-    fontSize: "1rem",
-    marginTop: "0.5rem",
-    marginBottom: "0.5rem",
-    marginLeft: "1rem",
-    color: "black",
-
-    "& input": {
-      marginRight: "0.5rem",
-    },
-  },
-  disableEndDate: {
-    backgroundColor: "f0f0f0",
-    color: "black",
-    width: "50%",
-    padding: "0.5rem",
-    margin: "0.5rem 0",
-    border: "1px solid #ddd",
-    borderRadius: "10px",
-  },
-});
 
 function DetailsContent() {
   const classes = useStyles();
@@ -169,21 +21,21 @@ function DetailsContent() {
   const [about, setAbout] = useState("");
   const [age, setAge] = useState("");
   const [hometown, setHometown] = useState("");
-  const [skills, setSkills] = useState([]); // Array of strings
-  const [skill, setSkill] = useState(""); // String
-  const [jobs, setJobs] = useState([]); // Array of strings
-  const [job, setJob] = useState(""); // String
-  const [jobname, setJobName] = useState(""); // String
-  const [company, setCompany] = useState(""); // String
-  const [startDate, setStartDate] = useState(""); // String
-  const [endDate, setEndDate] = useState(""); // String
-  const [jobDescription, setJobDescription] = useState(""); // String
+  const [skills, setSkills] = useState([]);
+  const [skill, setSkill] = useState("");
+  const [jobs, setJobs] = useState([]);
+  const [job, setJob] = useState("");
+  const [jobname, setJobName] = useState("");
+  const [company, setCompany] = useState("");
+  const [startDate, setStartDate] = useState("");
+  const [endDate, setEndDate] = useState("");
+  const [jobDescription, setJobDescription] = useState("");
   const [collectionType, setCollectionType] = useState("personal");
   const [linkedin, setLinkedin] = useState("");
   const [github, setGithub] = useState("");
-  const [ageFocused, setAgeFocused] = useState(false);
-  const details = useUserDetails(collectionType); // Fetch details based on collection type
+  const details = useUserDetails(collectionType);
   const [currentlyEmployed, setCurrentlyEmployed] = useState(false);
+  const [ageFocused, setAgeFocused] = useState(false);
 
   const handlePersonalDetailsSubmit = async (event) => {
     event.preventDefault();
@@ -324,6 +176,22 @@ function DetailsContent() {
       setEndDate(endDate);
     }
   };
+  const CallGeneratePDF = () => {
+    handleGeneratePDF({
+      name,
+      firstName,
+      lastName,
+      email,
+      phone,
+      age,
+      hometown,
+      about,
+      skills,
+      linkedin,
+      github,
+      jobs,
+    });
+  };
 
   return (
     <div className={classes.body}>
@@ -333,230 +201,81 @@ function DetailsContent() {
         into your templates
       </h3>
       <div className={classes.main}>
-        <form onSubmit={handlePersonalDetailsSubmit} className={classes.form}>
-          <div className={classes.titleimg}>
-            <h2>Personal Details </h2>
-            <img src={businessCard} alt="business card" />
-          </div>
-          <label>Name:</label>
-          <div className={classes.namesdiv}>
-            <input
-              className={`${classes.names} ${classes.input}`}
-              type="text"
-              placeholder="First Name"
-              value={firstName}
-              onChange={(e) => setFirstName(e.target.value)}
-              required
-            />
-            <input
-              className={`${classes.names} ${classes.input}`}
-              type="text"
-              placeholder="Last Name"
-              value={lastName}
-              onChange={(e) => setLastName(e.target.value)}
-              required
-            />
-          </div>
-          <br />
-          <label>Contact Details:</label>
-          <div className={classes.namesdiv}>
-            <input
-              type="email"
-              value={email}
-              placeholder="Email"
-              onChange={(e) => setEmail(e.target.value)}
-              required
-              className={classes.input}
-            />
-            <input
-              type="tel"
-              placeholder="Phone Number"
-              value={phone}
-              onChange={(e) => setPhone(e.target.value)}
-              required
-              className={classes.input}
-            />
-          </div>
-          <br />
-          <h3>About you:</h3>
-          <input
-            type={ageFocused ? "date" : "text"}
-            placeholder="Age"
-            value={age}
-            onChange={(e) => setAge(e.target.value)}
-            onFocus={() => setAgeFocused(true)}
-            onBlur={() => setAgeFocused(false)}
-            required
-            className={classes.input}
-          />
-          <input
-            type="text"
-            placeholder="Hometown"
-            value={hometown}
-            onChange={(e) => setHometown(e.target.value)}
-            required
-            className={classes.input}
-          />
-          <textarea
-            type="text"
-            placeholder="A little about yourself"
-            value={about}
-            onChange={(e) => setAbout(e.target.value)}
-            required
-            className={classes.input}
-          />
-          <br />
-          <div>
-            <h3>Skills:</h3>
-            <input
-              type="text"
-              placeholder="Type a skill"
-              value={skill}
-              onChange={(event) => setSkill(event.target.value)}
-              className={classes.skillinput}
-            />
-            <button type="submit" onClick={handleAddSkill}>
-              Add
-            </button>
-          </div>
-          <div className={classes.skills}>
-            {skills.map((skill) => (
-              <span key={skill}>
-                {skill}
-                <button
-                  className={classes.closebutton}
-                  onClick={() => handleDeleteSkill(skill)}
-                >
-                  x
-                </button>
-              </span>
-            ))}
-          </div>
-          <br />
-          <h3>Links:</h3>
-          <input
-            type="url"
-            placeholder="LinkedIn"
-            value={linkedin}
-            onChange={(e) => setLinkedin(e.target.value)}
-            className={classes.input}
-          />
-          <input
-            type="url"
-            placeholder="Github"
-            value={github}
-            onChange={(e) => setGithub(e.target.value)}
-            className={classes.input}
-          />
-          <button type="submit" className={classes.submitButton}>
-            Update
-          </button>
-        </form>
+        <PersonalDetailsForm
+          firstName={firstName}
+          setFirstName={setFirstName}
+          lastName={lastName}
+          setLastName={setLastName}
+          email={email}
+          setEmail={setEmail}
+          phone={phone}
+          setPhone={setPhone}
+          age={age}
+          setAge={setAge}
+          hometown={hometown}
+          setHometown={setHometown}
+          about={about}
+          setAbout={setAbout}
+          skills={skills}
+          setSkills={setSkills}
+          skill={skill}
+          setSkill={setSkill}
+          handleAddSkill={handleAddSkill}
+          handleDeleteSkill={handleDeleteSkill}
+          linkedin={linkedin}
+          setLinkedin={setLinkedin}
+          github={github}
+          setGithub={setGithub}
+          handlePersonalDetailsSubmit={handlePersonalDetailsSubmit}
+          classes={classes}
+          collectionType={collectionType}
+          setCollectionType={setCollectionType}
+          ageFocused={ageFocused}
+          setAgeFocused={setAgeFocused}
+        />
 
-        <form onSubmit={handleWorkHistorySubmit} className={classes.form}>
-          <div className={classes.titleimg}>
-            <h2>Work History</h2>
-            <img src={resume} alt="resume" />
-          </div>
-          <label>Job Details:</label>
-          <div className={classes.namesdiv}>
-            <input
-              className={`${classes.names} ${classes.input}`}
-              type="text"
-              placeholder="Job Title"
-              value={jobname}
-              onChange={(e) => setJobName(e.target.value)}
-              required
-            />
-            <input
-              className={`${classes.names} ${classes.input}`}
-              type="text"
-              placeholder="Company Name"
-              value={company}
-              onChange={(e) => setCompany(e.target.value)}
-              required
-            />
-          </div>
-          <br />
-          <label>Start and End Date:</label>
-          <div className={classes.namesdiv}>
-            <input
-              type="date"
-              value={startDate}
-              placeholder="Start Date"
-              onChange={(e) => setStartDate(e.target.value)}
-              required
-              className={classes.input}
-            />
-            <div className={classes.namesdiv}>
-              <label className={classes.checkboxLabel}>
-                <input
-                  type="checkbox"
-                  checked={currentlyEmployed}
-                  onChange={handleCurrentlyEmployedChange}
-                  className={classes.checkboxInput}
-                />
-                Current
-              </label>
-
-              {!currentlyEmployed ? (
-                <input
-                  type="date"
-                  placeholder="End Date"
-                  value={endDate}
-                  onChange={(e) => setEndDate(e.target.value)}
-                  required
-                  className={classes.input}
-                />
-              ) : (
-                <input
-                  type="date"
-                  placeholder="End Date"
-                  className={classes.disableEndDate}
-                  disabled
-                />
-              )}
-            </div>
-          </div>
-          <br />
-          <h3>About The Job:</h3>
-          <textarea
-            type="text"
-            placeholder="A little about the Job, your responsibilities, and what skills you learned"
-            value={jobDescription}
-            onChange={(e) => setJobDescription(e.target.value)}
-            required
-            className={classes.input}
-          />
-          <br />
-          <button
-            type="button"
-            onClick={handleAddJob}
-            className={classes.submitButton}
-          >
-            Add Job
-          </button>
-          <div className={classes.jobs}>
-            {jobs === undefined ? (
-              <p>No jobs yet.</p>
-            ) : (
-              jobs.map((job, index) => (
-                <div key={index}>
-                  <span>
-                    {job.job} at {job.company}
-                    <button
-                      className={classes.closebutton}
-                      onClick={() => handleDeleteJob(index)}
-                    >
-                      x
-                    </button>
-                  </span>
-                </div>
-              ))
-            )}
-          </div>
-        </form>
+        <WorkHistoryForm
+          jobs={jobs}
+          setJobs={setJobs}
+          job={job}
+          setJob={setJob}
+          jobname={jobname}
+          setJobName={setJobName}
+          company={company}
+          setCompany={setCompany}
+          startDate={startDate}
+          setStartDate={setStartDate}
+          endDate={endDate}
+          setEndDate={setEndDate}
+          jobDescription={jobDescription}
+          setJobDescription={setJobDescription}
+          handleAddJob={handleAddJob}
+          handleDeleteJob={handleDeleteJob}
+          handleWorkHistorySubmit={handleWorkHistorySubmit}
+          currentlyEmployed={currentlyEmployed}
+          setCurrentlyEmployed={setCurrentlyEmployed}
+          handleCurrentlyEmployedChange={handleCurrentlyEmployedChange}
+          classes={classes}
+        />
       </div>
+      <br />
+      {details != null ? (
+        <button className={classes.submitButton} style={{
+          width: "30%",
+        }} onClick={CallGeneratePDF}>
+          Generate PDF CV
+        </button>
+      ) : (
+        <button
+          style={{
+            backgroundColor: "grey",
+            color: "white",
+          }}
+          disabled
+        >
+          Generate PDF CV
+        </button>
+      )}
     </div>
   );
 }
